Move asset bar operator to new files
