+++ src/ui/operators/__init__.py
@@ -1,6 +1,7 @@
 # flake8: noqa
 """Blender UI Operators."""
 
+from .asset_bar import AssetBarOperator
 from .default_names import DefaultNamesOperator
 from .run_asset_bar import RunAssetBarWithContext
 from .transfer_data import TransferHana3DData
